requirements.txt
python-telegram-bot
nekos.py
requests
Pillow
pyyaml